Move chain test traits from chain_driver:: to chain:: module
