+++ crates/test/test-components/src/chain/mod.rs
@@ -1,0 +1,1 @@
+pub mod traits;
