+++ crates/test/test-components/src/chain/traits/mod.rs
@@ -1,0 +1,4 @@
+pub mod messages;
+pub mod queries;
+pub mod transfer;
+pub mod types;
