+++ crates/test/test-components/src/chain/traits/queries/mod.rs
@@ -1,0 +1,1 @@
+pub mod balance;
