+++ crates/test/test-components/src/chain/traits/transfer/local_transfer.rs
@@ -1,8 +1,8 @@
 use cgp_core::prelude::*;
 
-use crate::chain_driver::traits::types::address::HasAddressType;
-use crate::chain_driver::traits::types::amount::HasAmountType;
-use crate::chain_driver::traits::types::wallet::HasWalletType;
+use crate::chain::traits::types::address::HasAddressType;
+use crate::chain::traits::types::amount::HasAmountType;
+use crate::chain::traits::types::wallet::HasWalletType;
 
 #[derive_component(TokenLocalTransferrerComponent, TokenLocalTransferrer<Chain>)]
 #[async_trait]
