+++ crates/test/test-components/src/chain/traits/transfer/mod.rs
@@ -1,3 +1,2 @@
-pub mod balance;
 pub mod ibc_transfer;
 pub mod local_transfer;
