+++ crates/test/test-components/src/chain/traits/types/amount.rs
@@ -2,7 +2,7 @@ use core::fmt::{Debug, Display};
 
 use cgp_core::prelude::*;
 
-use crate::chain_driver::traits::types::denom::HasDenomType;
+use crate::chain::traits::types::denom::HasDenomType;
 
 pub type AmountOf<Chain> = <Chain as HasAmountType>::Amount;
