+++ crates/test/test-components/src/chain/traits/types/denom.rs
