+++ crates/test/test-components/src/chain/traits/types/mod.rs
@@ -1,0 +1,6 @@
+pub mod address;
+pub mod amount;
+pub mod denom;
+pub mod memo;
+pub mod tx_context;
+pub mod wallet;
