+++ crates/test/test-components/src/chain_driver/traits/mod.rs
@@ -1,6 +1,4 @@
 pub mod assert;
 pub mod build;
 pub mod fields;
-pub mod messages;
-pub mod queries;
 pub mod types;
