+++ crates/test/test-components/src/chain_driver/traits/types/mod.rs
@@ -1,7 +1,1 @@
-pub mod address;
-pub mod amount;
 pub mod chain;
-pub mod denom;
-pub mod memo;
-pub mod tx_context;
-pub mod wallet;
