+++ crates/test/test-components/src/lib.rs
@@ -5,6 +5,7 @@
 extern crate alloc;
 
 pub mod bootstrap;
+pub mod chain;
 pub mod chain_driver;
 pub mod driver;
 pub mod relay_driver;
